feat!: more improvements to eigen utilities and eigen context example wip
