+++ .gitmodules
@@ -24,11 +24,3 @@
 [submodule "blueprints/examples/contracts/lib/forge-std"]
 	path = blueprints/examples/contracts/lib/forge-std
 	url = https://github.com/foundry-rs/forge-std
-
-# For the example blueprint
-[submodule "blueprints/examples/contracts/lib/eigenlayer-middleware"]
-	path = blueprints/examples/contracts/lib/eigenlayer-middleware
-	url = https://github.com/Layr-Labs/eigenlayer-middleware
-[submodule "blueprints/examples/contracts/lib/forge-std"]
-	path = blueprints/examples/contracts/lib/forge-std
-	url = https://github.com/foundry-rs/forge-std
