+++ blueprints/examples/src/lib.rs
@@ -1,1 +1,4 @@
 pub mod examples;
+
+#[cfg(test)]
+mod tests;
