+++ blueprints/incredible-squaring-eigenlayer/src/lib.rs
@@ -21,3 +21,19 @@ load_abi!(
     INCREDIBLE_SQUARING_TASK_MANAGER_ABI_STRING,
     "contracts/out/IncredibleSquaringTaskManager.sol/IncredibleSquaringTaskManager.json"
 );
+
+sol!(
+    #[allow(missing_docs)]
+    #[sol(rpc)]
+    #[derive(Debug)]
+    PauserRegistry,
+    "./contracts/out/IPauserRegistry.sol/IPauserRegistry.json"
+);
+
+sol!(
+    #[allow(missing_docs, clippy::too_many_arguments)]
+    #[sol(rpc)]
+    #[derive(Debug)]
+    RegistryCoordinator,
+    "./contracts/out/RegistryCoordinator.sol/RegistryCoordinator.json"
+);
